requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.6.1
python-dotenv==1.0.0
lxml==5.1.0
pytz==2024.1